EVEREST-270 Return 400 instead of 502 if k8s cluster is unavailable
