+++ api/everest.go
@@ -23,19 +23,13 @@ import (
 	"fmt"
 	"io/fs"
 	"net/http"
-	"net/http/httputil"
-	"net/url"
-	"strings"
 	"sync"
 
-	"github.com/AlekSi/pointer"
 	"github.com/deepmap/oapi-codegen/pkg/middleware"
 	"github.com/labstack/echo/v4"
 	echomiddleware "github.com/labstack/echo/v4/middleware"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/percona/percona-everest-backend/cmd/config"
 	"github.com/percona/percona-everest-backend/model"
@@ -200,58 +194,4 @@ func (e *EverestServer) Shutdown(ctx context.Context) error {
 	case <-ctx.Done():
 		return ctx.Err()
 	}
-}
-
-func (e *EverestServer) proxyKubernetes(ctx echo.Context, kubernetesID, resourceName string) error {
-	cluster, err := e.storage.GetKubernetesCluster(ctx.Request().Context(), kubernetesID)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{
-			Message: pointer.ToString("Could not get a Kubernetes cluster"),
-		})
-	}
-	encodedSecret, err := e.secretsStorage.GetSecret(ctx.Request().Context(), kubernetesID)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{
-			Message: pointer.ToString("Could not retrieve kubeconfig"),
-		})
-	}
-
-	config, err := clientcmd.BuildConfigFromKubeconfigGetter("", newConfigGetter(encodedSecret).loadFromString)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusBadRequest, Error{
-			Message: pointer.ToString("Could not build kubeconfig"),
-		})
-	}
-	reverseProxy := httputil.NewSingleHostReverseProxy(
-		&url.URL{
-			Host:   strings.TrimPrefix(config.Host, "https://"),
-			Scheme: "https",
-		})
-	transport, err := rest.TransportFor(config)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusBadRequest, Error{
-			Message: pointer.ToString("Could not create REST transport"),
-		})
-	}
-	reverseProxy.Transport = transport
-	req := ctx.Request()
-	req.URL.Path = buildProxiedURL(ctx.Request().URL.Path, kubernetesID, resourceName, cluster.Namespace)
-	reverseProxy.ServeHTTP(ctx.Response(), req)
-	return nil
-}
-
-func buildProxiedURL(uri, kubernetesID, resourceName, namespace string) string {
-	// cut the /kubernetes part
-	uri = strings.TrimPrefix(uri, "/v1/kubernetes/"+kubernetesID)
-
-	// cut the resource name if present
-	uri = strings.TrimSuffix(uri, resourceName)
-
-	// remove kebab-case
-	uri = strings.ReplaceAll(uri, "-", "")
-	return fmt.Sprintf("/apis/everest.percona.com/v1alpha1/namespaces/%s%s%s", namespace, uri, resourceName)
 }
